Initial testing phase of receiveq
THINGS KIND OF WORK OMG
changed Makefile to create tap device on host
make sure to set up your interface, see ip(8) for more.

+++ kernel/virtio.h
@@ -122,7 +122,7 @@ struct virtio_net_hdr {
   uint16 gso_size;       // Maximum segment size (not including header).
   uint16 csum_start;       // The position to begin calculating the checksum.
   uint16 csum_offset;      // The position after ChecksumStart to store the checksum.
-  uint64 num_buffers;
+  uint16 num_buffers;
 };
